Refactor: Move story view code to be consistent
