+++ timesketch/views/sketch.py
@@ -44,6 +44,7 @@ from timesketch.models.sketch import SearchIndex
 from timesketch.models.sketch import SearchTemplate
 from timesketch.models.sketch import Timeline
 from timesketch.models.sketch import View
+from timesketch.models.story import Story
 from timesketch.models.user import Group
 from timesketch.models.user import User
 from timesketch.lib.datastores.elastic import ElasticsearchDataStore
@@ -254,53 +255,73 @@ def explore(sketch_id, view_id=None, searchtemplate_id=None):
 
 
 @sketch_views.route(
-    u'/sketch/<int:sketch_id>/explore/export/', methods=[u'GET'])
+    u'/sketch/<int:sketch_id>/graphs/', methods=[u'GET', u'POST'])
 @login_required
-def export(sketch_id):
-    """Generates CSV from search result.
+def graphs(sketch_id):
+    """Generates the sketch views template.
 
-    Args:
-        sketch_id: Primary key for a sketch.
     Returns:
-        CSV string with header.
+        Template with context.
     """
     sketch = Sketch.query.get_with_acl(sketch_id)
-    view = sketch.get_user_view(current_user)
-    query_filter = json.loads(view.query_filter)
-    query_dsl = json.loads(view.query_dsl)
-    indices = query_filter.get(u'indices', [])
+    graphs_enabled = current_app.config[u'GRAPH_BACKEND_ENABLED']
 
-    # Export more than the 500 first results.
-    max_events_to_fetch = 10000
-    query_filter[u'size'] = max_events_to_fetch
+    return render_template(
+        u'sketch/graphs.html',
+        sketch=sketch,
+        graphs_enabled=graphs_enabled)
+    
 
-    datastore = ElasticsearchDataStore(
-        host=current_app.config[u'ELASTIC_HOST'],
-        port=current_app.config[u'ELASTIC_PORT'])
+@sketch_views.route(u'/sketch/<int:sketch_id>/stories/')
+@sketch_views.route(u'/sketch/<int:sketch_id>/stories/<int:story_id>/')
+@login_required
+def story(sketch_id, story_id=None):
+    """Generates the story list template.
 
-    result = datastore.search(
-        sketch_id,
-        view.query_string,
-        query_filter,
-        query_dsl,
-        indices,
-        aggregations=None,
-        return_results=True)
+    Returns:
+        Template with context.
+    """
+    sketch = Sketch.query.get_with_acl(sketch_id)
+    graphs_enabled = current_app.config[u'GRAPH_BACKEND_ENABLED']
 
-    csv_out = StringIO()
-    csv_writer = csv.DictWriter(
-        csv_out,
-        fieldnames=[
-            u'timestamp', u'message', u'timestamp_desc', u'datetime',
-            u'timesketch_label', u'tag'
-        ])
-    csv_writer.writeheader()
-    for _event in result[u'hits'][u'hits']:
-        csv_writer.writerow(
-            dict((k, v.encode(u'utf-8') if isinstance(v, basestring) else v)
-                 for k, v in _event[u'_source'].iteritems()))
+    current_story = None
+    if story_id:
+        current_story = Story.query.get(story_id)
+    return render_template(
+        u'sketch/stories.html', sketch=sketch, story=current_story,
+        graphs_enabled=graphs_enabled)
 
-    return csv_out.getvalue()
+
+@sketch_views.route(
+    u'/sketch/<int:sketch_id>/views/', methods=[u'GET', u'POST'])
+@login_required
+def views(sketch_id):
+    """Generates the sketch views template.
+
+    Returns:
+        Template with context.
+    """
+    sketch = Sketch.query.get_with_acl(sketch_id)
+    trash_form = TrashViewForm()
+    graphs_enabled = current_app.config[u'GRAPH_BACKEND_ENABLED']
+
+    # Trash form POST
+    if trash_form.validate_on_submit():
+        if not sketch.has_permission(current_user, u'write'):
+            abort(HTTP_STATUS_CODE_FORBIDDEN)
+        view_id = trash_form.view_id.data
+        view = View.query.get(view_id)
+        # Check that this view belongs to the sketch
+        if view.sketch_id != sketch.id:
+            abort(HTTP_STATUS_CODE_NOT_FOUND)
+        view.set_status(status=u'deleted')
+        return redirect(u'/sketch/{0:d}/views/'.format(sketch.id))
+
+    return render_template(
+        u'sketch/views.html',
+        sketch=sketch,
+        trash_form=trash_form,
+        graphs_enabled=graphs_enabled)
 
 
 @sketch_views.route(
@@ -400,50 +421,50 @@ def timeline(sketch_id, timeline_id):
 
 
 @sketch_views.route(
-    u'/sketch/<int:sketch_id>/views/', methods=[u'GET', u'POST'])
+    u'/sketch/<int:sketch_id>/explore/export/', methods=[u'GET'])
 @login_required
-def views(sketch_id):
-    """Generates the sketch views template.
+def export(sketch_id):
+    """Generates CSV from search result.
 
+    Args:
+        sketch_id: Primary key for a sketch.
     Returns:
-        Template with context.
+        CSV string with header.
     """
     sketch = Sketch.query.get_with_acl(sketch_id)
-    trash_form = TrashViewForm()
-    graphs_enabled = current_app.config[u'GRAPH_BACKEND_ENABLED']
+    view = sketch.get_user_view(current_user)
+    query_filter = json.loads(view.query_filter)
+    query_dsl = json.loads(view.query_dsl)
+    indices = query_filter.get(u'indices', [])
 
-    # Trash form POST
-    if trash_form.validate_on_submit():
-        if not sketch.has_permission(current_user, u'write'):
-            abort(HTTP_STATUS_CODE_FORBIDDEN)
-        view_id = trash_form.view_id.data
-        view = View.query.get(view_id)
-        # Check that this view belongs to the sketch
-        if view.sketch_id != sketch.id:
-            abort(HTTP_STATUS_CODE_NOT_FOUND)
-        view.set_status(status=u'deleted')
-        return redirect(u'/sketch/{0:d}/views/'.format(sketch.id))
+    # Export more than the 500 first results.
+    max_events_to_fetch = 10000
+    query_filter[u'limit'] = max_events_to_fetch
 
-    return render_template(
-        u'sketch/views.html',
-        sketch=sketch,
-        trash_form=trash_form,
-        graphs_enabled=graphs_enabled)
+    datastore = ElasticsearchDataStore(
+        host=current_app.config[u'ELASTIC_HOST'],
+        port=current_app.config[u'ELASTIC_PORT'])
 
+    result = datastore.search(
+        sketch_id,
+        view.query_string,
+        query_filter,
+        query_dsl,
+        indices,
+        aggregations=None,
+        return_results=True)
 
-@sketch_views.route(
-    u'/sketch/<int:sketch_id>/graphs/', methods=[u'GET', u'POST'])
-@login_required
-def graphs(sketch_id):
-    """Generates the sketch views template.
+    csv_out = StringIO()
+    csv_writer = csv.DictWriter(
+        csv_out,
+        fieldnames=[
+            u'timestamp', u'message', u'timestamp_desc', u'datetime',
+            u'timesketch_label', u'tag'
+        ])
+    csv_writer.writeheader()
+    for _event in result[u'hits'][u'hits']:
+        csv_writer.writerow(
+            dict((k, v.encode(u'utf-8') if isinstance(v, basestring) else v)
+                 for k, v in _event[u'_source'].iteritems()))
 
-    Returns:
-        Template with context.
-    """
-    sketch = Sketch.query.get_with_acl(sketch_id)
-    graphs_enabled = current_app.config[u'GRAPH_BACKEND_ENABLED']
-
-    return render_template(
-        u'sketch/graphs.html',
-        sketch=sketch,
-        graphs_enabled=graphs_enabled)
+    return csv_out.getvalue()
